requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
boto3==1.35.70
python-dotenv==1.0.1
httpx==0.27.2
PyMuPDF==1.24.14
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
python-dateutil==2.9.0.post0
spacy>=3.7.0
